Reto8_TicTacToeOnline/back/serverOLD/requirements.txt
Django==1.9.4
django-filter==0.13.0
djangorestframework==3.3.3
drf-extensions==0.3.1
Markdown==2.6.6
wheel==0.24.0
Pillow==3.4.2
django-imagekit==3.3
pyfcm==1.1.3
